requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.116.1
gmqtt==0.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
